fix pom and package name
