+++ src/main/resources/io/jenkins/plugins/jfrog/ArtifactoryInstaller/config.jelly
