+++ src/main/resources/io/jenkins/plugins/jfrog/JfrogInstallation/config.jelly
