+++ src/main/resources/io/jenkins/plugins/jfrog/ReleasesInstaller/config.jelly
