--- src/test/java/jenkins/plugins/jfrog/JfrogInstallationTest.java
@@ -1,83 +1,0 @@
-package jenkins.plugins.jfrog;
-
-import hudson.FilePath;
-import hudson.model.Saveable;
-import hudson.tools.InstallSourceProperty;
-import hudson.tools.ToolProperty;
-import hudson.tools.ToolPropertyDescriptor;
-import hudson.util.DescribableList;
-import jenkins.model.Jenkins;
-import jenkins.plugins.jfrog.pipeline.PipelineTestBase;
-import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
-import org.jenkinsci.plugins.workflow.job.WorkflowJob;
-import org.jenkinsci.plugins.workflow.job.WorkflowRun;
-import org.jvnet.hudson.test.JenkinsRule;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-
-// TODO un-comment tests
-class JfrogInstallationTest extends PipelineTestBase {
-    // Jfrog CLI version which is accessible for all operating systems
-    public static final String jfrogCliTestVersion = "2.6.1";
-
-    /**
-     * Adds Jfrog cli tool as a global tool and verify installation.
-     * //     * @param jenkins Jenkins instance Injected automatically.
-     */
-//    @Test
-//    public void testJfrogCliInstallation(JenkinsRule jenkins) throws Exception{
-//        initPipelineTest(jenkins);
-//        configureJfrogCli();
-//        WorkflowRun job = runPipeline(jenkins, "");
-//        assertTrue(job.getLog().contains("jf version "+jfrogCliTestVersion));
-//        // remove, only for testing
-//        while (true) ;
-//    }
-    public static JfrogInstallation configureJfrogCli() throws IOException {
-        Saveable NOOP = () -> {
-        };
-        DescribableList<ToolProperty<?>, ToolPropertyDescriptor> r = new DescribableList<>(NOOP);
-        List<ReleasesInstaller> installers = new ArrayList<>();
-        installers.add(new ReleasesInstaller(jfrogCliTestVersion));
-        r.add(new InstallSourceProperty(installers));
-        JfrogInstallation jf = new JfrogInstallation("cli", "", r);
-        Jenkins.get().getDescriptorByType(JfrogInstallation.Descriptor.class).setInstallations(jf);
-        return jf;
-    }
-
-    /**
-     * Run pipeline script.
-     *
-     * @param name - Pipeline name from 'jenkins-artifactory-plugin/src/test/resources/integration/pipelines'
-     * @return the Jenkins job
-     */
-    WorkflowRun runPipeline(JenkinsRule jenkins, String name) throws Exception {
-        WorkflowJob project = jenkins.createProject(WorkflowJob.class);
-        FilePath slaveWs = slave.getWorkspaceFor(project);
-        if (slaveWs == null) {
-            throw new Exception("Slave workspace not found");
-        }
-        slaveWs.mkdirs();
-        String file = "pipeline {\n" +
-                "    agent any\n" +
-                "    tools {\n" +
-                "        \"jfrog\" \"cli\"\n" +
-                "    }\n" +
-                "    stages {\n" +
-                "        stage('Build') {\n" +
-                "            steps {\n" +
-                "                echo 'Building..'\n" +
-                "                sh 'jf -v'\n" +
-                "            }\n" +
-                "        }\n" +
-                "     }\n" +
-                "}";
-        project.setDefinition(new CpsFlowDefinition(file, false));
-        return jenkins.buildAndAssertSuccess(project);
-    }
-
-
-}
-
